+++ downhill-lib/CHANGELOG.md
@@ -4,6 +4,7 @@
 * `MetricTensor` is no longer required to be a `VectorSpace`
 * `T2`, `T3` pattern synonyms for `BVar`
 * `TangOptions`, `GradOptions` template haskell options.
+* `L2` metric.
 
 ## 0.1.0.0