+++ downhill-lib/CHANGELOG.md
@@ -3,9 +3,9 @@
 ## 0.2.0.0
 * `MetricTensor` is no longer required to be a `VectorSpace`
 * `T2`, `T3` pattern synonyms for `BVar`
-* `TangOptions`, `GradOptions` template haskell options.
-* `L2` metric.
-* Generics for `BasicVector`, `Dual`.
+* `L2` metric
+* Generics for `BasicVector`, `Dual`
+* Template Haskell scrapped
 
 ## 0.1.0.0